requirements.txt
Flask==1.1.2
requests==2.24.0
PyJWT==2.0.1